feat(v0.85.0): Assessment 13 medium findings — all 22 items implemented
Correctness (C13-02 – C13-11):
- C13-02: batch_decode raises PT512 on missing ID when strict_dictionary=on
- C13-03: blank-node-in-quoted-triple limitation documented + regression test
- C13-04: execute_drop/execute_clear mutation journal doc comment added
- C13-05: plan cache INFERENCE_MODE normalised with trim+lowercase
- C13-06: GRAPH ?g default-graph exclusion documented + regression test
- C13-07: batch_decode guard tightened to id==0 (was id<=0)
- C13-08: encode_token in magic.rs detects typed literals (^^< suffix)
- C13-09: parse_nt_triple rejects IRIs >4KiB; requires trailing >
- C13-10: xsd:dateTime sub-ms truncation documented + regression test
- C13-11: describe_cbd depth capped by pg_ripple.describe_max_depth GUC (default 16)

Performance (P13-02 – P13-08):
- P13-02: encode_batch() single-CTE API + batch_encode_terms() SQL function
- P13-03: merge-worker heartb- P13-03: merge-worker heartb- P13-03: merge-worker heartb- P13-03: merge-worker heartb- P13-03: merge-worker heartb- P13-03: merge-og rule streaming in batches of 100
- P13-06: partition_into_parallel_groups() DFS cycle pre-check
- P13-07: PathCtx.counter private; next_alias()/value() accessors
- P13-08: dictionary_hot_cache_hits/misses Prometheus counters

Code Quality (Q13-02 – Q13-05):
- Q13-02: src/schema.rs split into schema/{tables,views,triggers,r- Q13-02: src/schema.rs split into schema/{tables,views,triggen/{circuit,policy,http,decode}.rs
- Q13-04: CI lint-file-size job (>18- Q13-04: CI lint-file-size job (>18- Q13-04: CI lint-file-size jead_code)] markers audited with justification comments

Concurrency (CC1Concurrency (CC1Concurrency (CC1Concurrency (CC1Concurrency (gkill.sh Concurrency (CC1Concurrency (CC1Concurrency (CC1Concurrency (CC1Concurrency (000 namespace)

Version: 0.84.0 → 0.85.0
Migration: sql/pg_ripple--0.84.0--0.85.0.sql (no DDL changes)
