+++ roadmap/v0.85.0.md
@@ -2,7 +2,7 @@
 
 > **Full technical details:** [v0.85.0-full.md](v0.85.0-full.md)
 
-**Status: Planned** | **Scope: Large**
+**Status: Released** | **Scope: Large**
 
 > v0.85.0 closes the 22 Medium-priority correctness, performance, code-quality,
 > and concurrency findings from Assessment 13
@@ -17,70 +17,70 @@
 ## Headline outcomes
 
 ### Correctness (C13-02 – C13-11)
-- [ ] **C13-02** — `batch_decode` errors with `PT512` when an ID is missing
+- [x] **C13-02** — `batch_decode` errors with `PT512` when an ID is missing
   from the dictionary and `pg_ripple.strict_dictionary = on` (currently returns
   silent empty string)
-- [ ] **C13-03** — blank-node-in-quoted-triple limitation documented in
+- [x] **C13-03** — blank-node-in-quoted-triple limitation documented in
   `docs/src/reference/sparql-compliance.md`; round-trip regression test added;
   input rejected with `PT512` or canonicalised
-- [ ] **C13-04** — `execute_drop` / `execute_clear`: comment in code documents
+- [x] **C13-04** — `execute_drop` / `execute_clear`: comment in code documents
   that the caller must flush the mutation journal before calling; or add
   explicit `record_clear()` + `flush()` inside
-- [ ] **C13-05** — plan cache key trims whitespace and lowercases
+- [x] **C13-05** — plan cache key trims whitespace and lowercases
   `INFERENCE_MODE` value before hashing, preventing cache misses due to
   capitalisation or padding
-- [ ] **C13-06** — `GRAPH ?g` default-graph exclusion behaviour documented in
+- [x] **C13-06** — `GRAPH ?g` default-graph exclusion behaviour documented in
   `docs/src/reference/sparql-compliance.md`; regression test added asserting
   current behaviour
-- [ ] **C13-07** — `batch_decode` warning guard tightened to `id == 0` (was
+- [x] **C13-07** — `batch_decode` warning guard tightened to `id == 0` (was
   `id <= 0`; negative IDs should not occur after v0.81.0 dict-subxact fix but
   were previously silently passed)
-- [ ] **C13-08** — `encode_token` in `src/datalog/magic.rs` detects typed
+- [x] **C13-08** — `encode_token` in `src/datalog/magic.rs` detects typed
   literals (`^^<` suffix) and routes to `encode_typed_literal()` instead of
   plain string encoding
-- [ ] **C13-09** — `parse_nt_triple` adds bounds check: reject IRIs longer than
+- [x] **C13-09** — `parse_nt_triple` adds bounds check: reject IRIs longer than
   4 KiB; require trailing `>` for IRI tokens
-- [ ] **C13-10** — `xsd:dateTime` sub-millisecond precision limit (3 decimal
+- [x] **C13-10** — `xsd:dateTime` sub-millisecond precision limit (3 decimal
   places) documented in reference docs; round-trip regression test at 9
   decimal places asserts truncation to 3
-- [ ] **C13-11** — `describe_cbd` recursion depth capped via
+- [x] **C13-11** — `describe_cbd` recursion depth capped via
   `pg_ripple.describe_max_depth` GUC (default 16); prevents runaway recursion
   on cyclic or deep graphs
 
 ### Performance (P13-02 – P13-08)
-- [ ] **P13-02** — `encode_batch` API: single CTE inserts many terms and
+- [x] **P13-02** — `encode_batch` API: single CTE inserts many terms and
   returns all IDs in one SPI round-trip, reducing per-triple SPI overhead in
   bulk load paths
-- [ ] **P13-03** — merge-worker heartbeat log throttled to once per 60 seconds;
+- [x] **P13-03** — merge-worker heartbeat log throttled to once per 60 seconds;
   `_pg_ripple.merge_worker_status` table still updated every cycle
-- [ ] **P13-04** — `SET LOCAL` calls batched into a single SPI round-trip in
+- [x] **P13-04** — `SET LOCAL` calls batched into a single SPI round-trip in
   `execute_select()`
-- [ ] **P13-05** — Datalog rule streaming in batches of 100 rules, reducing
+- [x] **P13-05** — Datalog rule streaming in batches of 100 rules, reducing
   peak memory usage for large rule sets
-- [ ] **P13-06** — `partition_into_parallel_groups()` pre-checks for cycles in
+- [x] **P13-06** — `partition_into_parallel_groups()` pre-checks for cycles in
   the intra-stratum head-group dependency graph before SCC evaluation
-- [ ] **P13-07** — `PathCtx.counter` field made private; `next_alias()` method
+- [x] **P13-07** — `PathCtx.counter` field made private; `next_alias()` method
   exposed as the sole mutation point
-- [ ] **P13-08** — `dictionary_hot_cache_hits_total` and
+- [x] **P13-08** — `dictionary_hot_cache_hits_total` and
   `dictionary_hot_cache_misses_total` Prometheus counters added
 
 ### Code quality (Q13-02 – Q13-05)
-- [ ] **Q13-02** — `src/schema.rs` (1,939 lines) split into `schema/tables.rs`,
+- [x] **Q13-02** — `src/schema.rs` (1,939 lines) split into `schema/tables.rs`,
   `schema/views.rs`, `schema/triggers.rs`, `schema/rls.rs`
-- [ ] **Q13-03** — `src/sparql/federation.rs` (1,693 lines) split into
+- [x] **Q13-03** — `src/sparql/federation.rs` (1,693 lines) split into
   `federation/circuit.rs`, `federation/policy.rs`, `federation/http.rs`,
   `federation/decode.rs`
-- [ ] **Q13-04** — CI lint gate: any `src/**/*.rs` file exceeding 1,800 lines
+- [x] **Q13-04** — CI lint gate: any `src/**/*.rs` file exceeding 1,800 lines
   fails the build unless the file contains an
   `// @allow-large-file: <reason>` comment
-- [ ] **Q13-05** — all 36 `#[allow(dead_code)]` markers audited; each either
+- [x] **Q13-05** — all 36 `#[allow(dead_code)]` markers audited; each either
   wired up or deleted
 
 ### Concurrency (CC13-01 – CC13-02)
-- [ ] **CC13-01** — VP-promotion crash-recovery regression test:
+- [x] **CC13-01** — VP-promotion crash-recovery regression test:
   `tests/crash_recovery/promote_sigkill.sh` SIGKILLs the backend between the
   advisory-lock acquisition and the CTE commit; asserts predicate is promoted
   correctly on restart via `recover_interrupted_promotions()`
-- [ ] **CC13-02** — merge fence lock scoped per-predicate (advisory lock keyed
+- [x] **CC13-02** — merge fence lock scoped per-predicate (advisory lock keyed
   on `hash(predicate_id)`); global lock `0x5052_5000` reserved for Citus
   rebalance events only
